# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: G Creator Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auctions:
    get:
      operationId: listAuctions
      tags: [auctions]
      summary: List / search auctions
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: itemType
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [ending_soon, highest_bid, newest, buy_now] }
        - in: query
          name: limit
          schema: { type: integer, default: 24 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: sellerId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Auction"
    post:
      operationId: createAuction
      tags: [auctions]
      summary: Create an auction listing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /auctions/{auctionId}:
    get:
      operationId: getAuction
      tags: [auctions]
      summary: Get a single auction
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "404":
          description: Not found
    patch:
      operationId: updateAuction
      tags: [auctions]
      summary: Update your auction (owner only, while active)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuctionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "403":
          description: Forbidden
    delete:
      operationId: deleteAuction
      tags: [auctions]
      summary: Cancel / delete your auction (no bids only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auctions/{auctionId}/bids:
    get:
      operationId: listAuctionBids
      tags: [auctions]
      summary: Get bid history for an auction
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionBidEntry"
    post:
      operationId: placeBid
      tags: [auctions]
      summary: Place a bid on an auction
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidBody"
      responses:
        "200":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          description: Bid too low or auction ended
        "401":
          description: Unauthorized
  /auctions/{auctionId}/buy-now:
    post:
      operationId: buyNow
      tags: [auctions]
      summary: Instantly purchase an auction at the buy-now price
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          description: No buy-now price or already ended
  /auctions/{auctionId}/watch:
    post:
      operationId: toggleAuctionWatch
      tags: [auctions]
      summary: Toggle watching an auction
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchToggleResponse"
  /auctions/{auctionId}/end:
    post:
      operationId: endAuction
      tags: [auctions]
      summary: Manually end your own auction
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"

  # MERCH
...
components:
  schemas:
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        accountTier:
          type: string
          enum: [free, creator, pro, elite]
          default: free
        idVerificationStatus:
          type: string
          enum: [none, pending, approved, rejected]
          default: none
        isFollowing:
          type: boolean
        followersCount:
          type: integer
...
    Auction:
      type: object
      required: [id, sellerId, title, itemType, condition, startingBid, bidCount, watchCount, status, endTime, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
          nullable: true
        condition:
          type: string
          enum: [new, like_new, used, collectible]
        itemType:
          type: string
          enum: [physical, digital, experience, collectible, commission, ticket]
        startingBid:
          type: number
        reservePrice:
          type: ["number", "null"]
        buyNowPrice:
          type: ["number", "null"]
        currentBid:
          type: ["number", "null"]
        currentBidderId:
          type: ["integer", "null"]
        currentBidder:
          $ref: "#/components/schemas/UserSummary"
        bidCount:
          type: integer
        watchCount:
          type: integer
        status:
          type: string
          enum: [active, ended, cancelled, sold]
        shippingInfo:
          type: ["string", "null"]
        endTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        seller:
          $ref: "#/components/schemas/UserSummary"
        isWatching:
          type: boolean
    AuctionBidEntry:
      type: object
      required: [id, auctionId, bidderId, amount, createdAt]
      properties:
        id:
          type: integer
        auctionId:
          type: integer
        bidderId:
          type: integer
        amount:
          type: number
        isWinning:
          type: boolean
        createdAt:
          type: string
          format: date-time
        bidder:
          $ref: "#/components/schemas/UserSummary"
    CreateAuctionBody:
      type: object
      required: [title, startingBid, endTime, itemType, condition]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        condition:
          type: string
          enum: [new, like_new, used, collectible]
        itemType:
          type: string
          enum: [physical, digital, experience, collectible, commission, ticket]
        startingBid:
          type: number
          minimum: 0.01
        reservePrice:
          type: number
        buyNowPrice:
          type: number
        shippingInfo:
          type: string
        endTime:
          type: string
          format: date-time
    UpdateAuctionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        buyNowPrice:
          type: number
        shippingInfo:
          type: string
    PlaceBidBody:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 0.01
    WatchToggleResponse:
      type: object
      required: [watching]
      properties:
        watching:
          type: boolean
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...